moved test helper functions and variables into utils file
