+++ tests/ingredients.test.ts
@@ -1,71 +1,6 @@
-import { setupDb } from './utils.js';
+import { createSecondaryUser, setupDb, signUpAndCreateRecipe, testIngredient } from './utils.js';
 import request from 'supertest';
-import app from '../lib/app.js';
-import { UserService } from '../lib/services/UserService.js';
 import { Ingredient } from '../lib/models/Ingredient.js';
-
-const testUser = {
-  email: 'test@user.com',
-  password: '123456',
-  username: 'test_user'
-};
-
-const testUser2 = {
-  email: 'test2@user.com',
-  password: 'password',
-  username: 'second_user'
-};
-
-const testRecipe = {
-  name: 'mac and cheese',
-  description: 'so cheesy and delicious'
-};
-
-const testIngredient = {
-  name: 'milk',
-  amount: '1 cup'
-};
-
-type RecipeAgentData = {
-  agent: request.SuperAgentTest;
-  token: string;
-  userId: string;
-  recipeId: string;
-}
-
-type SecondUserData = {
-  token2: string;
-  secondUserId: string;
-}
-
-
-async function signUpAndCreateRecipe(): Promise<RecipeAgentData> {
-  const agent = request.agent(app);
-
-  const signUpRes = await agent.post('/users').send(testUser);
-  const { token } = signUpRes.body;
-
-  const userRes = await agent.get('/users/me')
-    .set('Authorization', `Bearer ${token}`);
-  const userId = userRes.body.user.id;
-
-  const newRecipeRes = await agent.post('/recipes')
-    .set('Authorization', `Bearer ${token}`)
-    .send(testRecipe);
-  const recipeId = newRecipeRes.body.recipe.id;
-
-  return { agent, token, userId, recipeId };
-}
-
-async function createSecondaryUser(agent: request.SuperAgentTest): Promise<SecondUserData> {
-  const secondUser = await UserService.create(testUser2);
-  const secondUserId = secondUser.id;
-
-  const signUpRes = await agent.post('/users/sessions').send(testUser2);
-  const { token: token2 } = signUpRes.body;
-
-  return { token2, secondUserId };
-}
 
 describe('POST /ingredients tests', () => {
   beforeEach(setupDb);
