+++ tests/recipe-shares.test.ts
@@ -1,74 +1,6 @@
 /* @jest-environment node */
-import { setupDb } from './utils.js';
-import request from 'supertest';
-import app from '../lib/app.js';
+import { setupDb, signUpAndCreateRecipe, signUpAndShareRecipe, testRecipe, testUser2, testUser3 } from './utils.js';
 import { UserService } from '../lib/services/UserService.js';
-
-const testUser = {
-  email: 'test@user.com',
-  password: '123456',
-  username: 'test_user'
-};
-
-const testUser2 = {
-  email: 'test2@user.com',
-  password: 'password',
-  username: 'second_user'
-};
-
-const testUser3 = {
-  email: 'third@user.com',
-  password: 'password',
-  username: 'third_user'
-};
-
-const testRecipe = {
-  name: 'mac and cheese',
-  description: 'so cheesy and delicious'
-};
-
-interface RecipeAgentData {
-  agent: request.SuperAgentTest;
-  token: string;
-  recipeId: string;
-}
-
-interface SharedRecipeAgentData extends RecipeAgentData {
-  token2: string;
-  shareId: string;
-  sharedUserId: string;
-}
-
-async function signUpAndCreateRecipe(): Promise<RecipeAgentData> {
-  const agent = request.agent(app);
-
-  const signUpRes = await agent.post('/users').send(testUser);
-  const { token } = signUpRes.body;
-
-  const recipeRes = await agent.post('/recipes')
-    .set('Authorization', `Bearer ${token}`)
-    .send(testRecipe);
-  const recipeId = recipeRes.body.recipe.id;
-
-  return { agent, token, recipeId };
-}
-
-async function signUpAndShareRecipe(): Promise<SharedRecipeAgentData> {
-  const { agent, token, recipeId } = await signUpAndCreateRecipe();
-
-  const secondUser = await UserService.create(testUser2);
-  const sharedUserId = secondUser.id;
-
-  const shareRes = await agent.post('/recipe-shares')
-    .set('Authorization', `Bearer ${token}`)
-    .send({ recipeId, userId: secondUser.id, editable: false });
-  const shareId = shareRes.body.recipeShare.id;
-
-  const { token: token2 } = (await agent.post('/users/sessions')
-    .send(testUser2)).body;
-
-  return { agent, token, token2, recipeId, shareId, sharedUserId };
-}
 
 
 describe('POST /recipe-shares tests', () => {
