+++ tests/recipe-steps.test.ts
@@ -1,47 +1,4 @@
-import { setupDb } from './utils.js';
-import request from 'supertest';
-import app from '../lib/app.js';
-
-const testUser = {
-  email: 'test@user.com',
-  password: '123456',
-  username: 'test_user'
-};
-
-const testRecipe = {
-  name: 'mac and cheese',
-  description: 'so cheesy and delicious'
-};
-
-const testStep = {
-  num: 1,
-  detail: 'boil the pasta'
-};
-
-type RecipeAgentData = {
-  agent: request.SuperAgentTest;
-  token: string;
-  userId: string;
-  recipeId: string;
-}
-
-async function signUpAndCreateRecipe(): Promise<RecipeAgentData> {
-  const agent = request.agent(app);
-
-  const signUpRes = await agent.post('/users').send(testUser);
-  const { token } = signUpRes.body;
-
-  const userRes = await agent.get('/users/me')
-    .set('Authorization', `Bearer ${token}`);
-  const userId = userRes.body.user.id;
-
-  const newRecipeRes = await agent.post('/recipes')
-    .set('Authorization', `Bearer ${token}`)
-    .send(testRecipe);
-  const recipeId = newRecipeRes.body.recipe.id;
-
-  return { agent, token, userId, recipeId };
-}
+import { setupDb, signUpAndCreateRecipe, testStep } from './utils.js';
 
 describe('POST /recipe-steps', () => {
   beforeEach(setupDb);
